+++ src/index.ts
@@ -1,5 +1,7 @@
 import usePython from "./hooks/use-python";
 import usePythonConsole from "./hooks/use-python-console";
 import { PythonProvider } from "./providers/python-provider";
+import type { Packages } from "./types";
 
 export { PythonProvider, usePython, usePythonConsole };
+export type { Packages };