+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # solid-highlight
 
+## 0.0.9
+
+### Patch Changes
+
+- fix package json
+
+## 0.0.8
+
+### Patch Changes
+
+- bundle workers
+
+## 0.0.7
+
+### Patch Changes
+
+- remove ts file extensions
+
+## 0.0.6
+
+### Patch Changes
+
+- add ts file extension to console worker
+
+## 0.0.5
+
+### Patch Changes
+
+- inline worker url
+
+## 0.0.4
+
+### Patch Changes
+
+- at ts file extension to python-worker file
+
+## 0.0.3
+
+### Patch Changes
+
+- 931d761: remove solid export
+- 931d761: remove solid export
+- expose packages type
+
 ## 0.0.2
 
 ### Patch Changes